requirements.txt
imutils==0.5.4
numpy==1.23.5
opencv_python==4.9.0.80
Pillow==10.4.0
streamlit==1.34.0
tensorflow==2.16.1
keras==2.16.0
scikit-learn==1.2.2
matplotlib==3.8.0
opencv-python-headless==4.8.0.76
numpy==1.25.2
pandas==2.0.2

